+++ libs/ngx-editorjs/src/lib/config/module-config.ts
@@ -1,11 +1,11 @@
-import { NgxEditorJSConfig } from '../types/config';
+import { NgxEditorJSModuleConfig } from '../types/config';
 import { createEditorJSConfig } from './editor-config';
 
 /**
  * Creates the configuration from the module config
  * @param config Optional module configurations
  */
-export function createModuleConfig(config?: NgxEditorJSConfig): NgxEditorJSConfig {
+export function createModuleConfig(config?: NgxEditorJSModuleConfig): NgxEditorJSModuleConfig {
   if (!config || !config.editorjs) {
     return {
       editorjs: createEditorJSConfig()