+++ libs/ngx-editorjs/src/lib/types/editorjs-service.ts
@@ -1,45 +1,4 @@
 import { EditorConfig } from '@editorjs/editorjs';
-/**
- * Interface describing an action to trigger EditorJS
- */
-export interface EditorJSAction<T = any> {
-  /**
-   * The name of the action to trigger
-   */
-  action: string;
-
-  /**
-   * The payload for the action <T>
-   */
-  payload?: T;
-}
-
-/**
- * Supported actions by the API
- */
-export enum EditorJSActionTypes {
-  /**
-   * Create a new editor
-   */
-  CreateEditor = 'createEditor',
-
-  /**
-   * Update an editor instance
-   */
-  UpdateEditor = 'updateEditor',
-  /**
-   * Clear a editor instance
-   */
-  ClearEditor = 'clearEditor',
-  /**
-   * Destroy an Editor instance
-   */
-  DestroyEditor = 'destroyEditor',
-  /**
-   * Save the blocks in an editor instance
-   */
-  SaveEditor = 'saveEditor'
-}
 
 /**
  * EditorJS options